Handle wall-drawing more correctly.
But what a mess! Going to replace this with a better option soon.

basically: I'm drawing each tile with a background and its walls.

A wall's presence on one tile's (tile X's) north side indicates that
the tile to its north (tile N) has a wall on its south side. These two
look like a single thick wall when on their own.

Imagine if Tile X also has a wall on its west side. The tile to the
west of X (let's say, tile W) has its own corresponding wall on its
east side, creating another seemingly single thick wall.

But if Tile N does not have a wall on its west, the corner of the
walls will look strange to the player:

                    --- south wall of N ---
                   |+-- north wall of X ---
                   ||
east wall of W ->  || <- west wall of X
                   ||

There is an apparent gap in the corner of the wall! That pixel is
actually a part of tile NW, to the northwest of tile X, which doesn't
have any walls.

So, there's a lot of ways to fix this, but they boil down to:

 1 - only have one wall rather than double-walls
 2 - extend the walls so they'd overlap better

Option 1 introduces a lot of tiles-reading-map-state during drawing,
which sounds pretty icky. Option 2 has a pitfall though: the background
of tile NW being drawn can overwrite the walls drawn for tiles N, W, or
X. So, I need to modify the times that the background is drawn versus
the walls, and here I've just brute-force stopped the walls from being
overwritten by drawing the background far less frequently and drawing
the walls, comparatively, *a lot*.