Reimagine maze walls, add enemies, add health.
The joy of side projects -- sane commit size and granularity can be
thrown to the wind :)

This commit reworks tiles and walls so that all walls are whole tiles,
adds enemies, and adds health for enemies and players with some cheap
attack effects and entity-scaling for a given health.

It also drops a few particularly bad looking palettes, reworks the way
the game's level transitions function, and changes the counts of things
in the levels.