Fix end-of-game map-change bug.
Enemies already in the update loop were attempting to perform their
work on a map that no longer existed, which could lead to out of bounds
checks. The hacky way I've fixed this for now is to use a global 'break
out of the update loop' flag, but a likely-better way I might use in
the future is to find some smart way to defer a map change until all
updates are complete for a given timestep.