Much prettier fog of war effect.
Each cell oscillates through a range of colours, and is broken into
subcells that themselves oscillate around that colour at a random
frequency, with color variations depending on the content of the map.

Also resize and change the number of maps, number of coins, enemies,
etc, given the new camera features and more information being shown
on the map to the observant user.